requirements.txt
torch>=2.0.0
mamba-ssm>=1.0.1
transformers>=4.30.0
datasets>=2.12.0
wandb>=0.15.0
pytest>=7.0.0
pyyaml>=6.0
tqdm>=4.65.0
einops>=0.6.1
